src/requirements.txt
Flask==2.0.2
numpy==1.21.4
Pillow==8.4.0
scikit_learn==1.0.1
